Added a `jest.setup.ts`, which generates the report.
