+++ jest.config.cjs
@@ -1,6 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
-  testPathIgnorePatterns: [
-    './dist/',
-  ],
+  testPathIgnorePatterns: ['./dist/'],
+  setupFiles: ['<rootDir>/jest.setup.ts'],
 };
