+++ src/tests/backwards-compatibility/backwards-compatibility.test.ts
@@ -1,11 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 
-import {
-  Extractor,
-  ExtractorConfig,
-  ExtractorResult,
-} from '@microsoft/api-extractor';
+
 import {
   ApiClass,
   ApiConstructor,
@@ -18,32 +14,7 @@ import {
   Parameter
 } from '@microsoft/api-extractor-model';
 
-describe('Generate API report', () => {
-  it('should generate an api report', () => {
-    const apiExtractorJsonPath: string = path.join(
-      __dirname,
-      'api-extractor.json'
-    );
-    const extractorConfig: ExtractorConfig =
-      ExtractorConfig.loadFileAndPrepare(apiExtractorJsonPath);
 
-    const extractorResult: ExtractorResult = Extractor.invoke(extractorConfig, {
-      localBuild: true,
-      showVerboseMessages: false,
-    });
-
-    if (extractorResult.succeeded) {
-      console.log(`API Extractor completed successfully`);
-      process.exitCode = 0;
-    } else {
-      console.error(
-        `API Extractor completed with ${extractorResult.errorCount} errors` +
-          ` and ${extractorResult.warningCount} warnings`
-      );
-      process.exitCode = 1;
-    }
-  });
-});
 
 describe('Validate API report', () => {
   let failure = false;
